cppad: move in dedicated file operations
